+++ src/main/java/com/basketbandit/core/BasketBandit.java
@@ -127,6 +127,7 @@ class BasketBandit extends ListenerAdapter {
             settingsList.add("commandprefix");
             settingsList.add("deleteexecuted");
             settingsList.add("commandlogging");
+            settingsList.add("announcenowplaying");
 
             Cache.STANDARD_STRINGS = new String[1];
             Cache.STANDARD_STRINGS[0] = Configuration.VERSION;