+++ src/main/java/com/basketbandit/core/modules/audio/commands/CommandCurrent.java
@@ -32,7 +32,6 @@ public class CommandCurrent extends Command {
                     .addField("Duration", Utils.getTimestamp(track.getPosition()) + "/" + Utils.getTimestamp(track.getDuration()), true)
                     .addField("Channel", track.getInfo().author, true)
                     .setFooter(Configuration.VERSION + " · Requested by " + e.getMember().getEffectiveName(), e.getGuild().getMemberById(Configuration.BOT_ID).getUser().getAvatarUrl());
-
             MessageHandler.sendMessage(e, queuedTrack.build());
         } else {
             EmbedBuilder embed = new EmbedBuilder().setTitle("There isn't a track currently playing.");