+++ src/main/java/com/basketbandit/core/modules/core/commands/CommandSettings.java
@@ -5,9 +5,8 @@ import com.basketbandit.core.Configuration;
 import com.basketbandit.core.database.DatabaseConnection;
 import com.basketbandit.core.database.DatabaseFunctions;
 import com.basketbandit.core.modules.Command;
-import com.basketbandit.core.modules.core.settings.SettingCommandLogging;
 import com.basketbandit.core.modules.core.settings.SettingCommandPrefix;
-import com.basketbandit.core.modules.core.settings.SettingDeleteExecuted;
+import com.basketbandit.core.modules.core.settings.SettingExecuteBoolean;
 import com.basketbandit.core.utils.MessageHandler;
 import com.basketbandit.core.utils.Sanitise;
 import com.basketbandit.core.utils.Utils;
@@ -32,7 +31,7 @@ public class CommandSettings extends Command {
 
                 // Check to make sure the command is a valid command.
                 if(!Cache.SETTINGS.contains(commandParameters[0].toLowerCase())) {
-                    EmbedBuilder embed = new EmbedBuilder().setTitle(commandParameters[1].toUpperCase() + "_ is not a valid setting.");
+                    EmbedBuilder embed = new EmbedBuilder().setTitle("_" + commandParameters[1].toUpperCase() + "_ is not a valid setting.");
                     MessageHandler.sendMessage(e, embed.build());
                     return;
                 }
@@ -41,28 +40,18 @@ public class CommandSettings extends Command {
                     return;
                 }
 
-                if(commandParameters[0].equalsIgnoreCase("deleteExecuted")) {
+                if(commandParameters[0].equalsIgnoreCase("commandLogging") || commandParameters[0].equalsIgnoreCase("deleteExecuted") || commandParameters[0].equalsIgnoreCase("announceNowPlaying")) {
                     if(!commandParameters[1].equalsIgnoreCase("true") && !commandParameters[1].equalsIgnoreCase("false")) {
-                        EmbedBuilder embed = new EmbedBuilder().setTitle(commandParameters[1].toUpperCase() + "_ is not a valid value. (Valid: TRUE, FALSE)");
+                        EmbedBuilder embed = new EmbedBuilder().setTitle("_" + commandParameters[1].toUpperCase() + "_ is not a valid value. (Valid: TRUE, FALSE)");
                         MessageHandler.sendMessage(e, embed.build());
                         return;
                     }
-                    new SettingDeleteExecuted(e, commandParameters[1]);
+                    new SettingExecuteBoolean(e, commandParameters[0].toLowerCase(), commandParameters[1]);
                     return;
                 }
 
                 if(commandParameters[0].equalsIgnoreCase("commandPrefix")) {
                     new SettingCommandPrefix(e, commandParameters[1]);
-                    return;
-                }
-
-                if(commandParameters[0].equalsIgnoreCase("commandLogging")) {
-                    if(!commandParameters[1].equalsIgnoreCase("true") && !commandParameters[1].equalsIgnoreCase("false")) {
-                        EmbedBuilder embed = new EmbedBuilder().setTitle(commandParameters[1].toUpperCase() + "_ is not a valid value. (Valid: TRUE, FALSE)");
-                        MessageHandler.sendMessage(e, embed.build());
-                        return;
-                    }
-                    new SettingCommandLogging(e, commandParameters[1]);
                 }
 
             } else {
@@ -72,12 +61,14 @@ public class CommandSettings extends Command {
                     ResultSet resultSet = new DatabaseFunctions().getServerSettings(connection, e.getGuild().getId());
                     resultSet.next();
 
+                    // Embed displaying all of the current settings for the server, giving information about each setting.
                     EmbedBuilder commandModules = new EmbedBuilder()
                         .setTitle("Settings for  + e.getGuild().getName() + ")
                         .setDescription("Settings can be changed by typing '<prefix>settings [setting] [value]' where [setting] is a value found below and [value] is a valid value, with special values like booleans being either TRUE or FALSE (case insensitive)")
                             .addField("commandPrefix", "[" + resultSet.getString("commandPrefix") + "] - The message prefix used to symbolise a command.", false)
                             .addField("deleteExecuted", "[" + resultSet.getBoolean("deleteExecuted") + "] - Deletes the users command string when it is executed.", false)
                             .addField("commandLogging", "[" + resultSet.getBoolean("commandLogging") + "] - Sends executed commands to a predefined logging channel.", false)
+                            .addField("announceNowPlaying", "[" + resultSet.getBoolean("announceNowPlaying") + "] - Sends information of the current track when it changes.", false)
                         .setFooter(Configuration.VERSION, e.getGuild().getMemberById(Configuration.BOT_ID).getUser().getAvatarUrl());
                     MessageHandler.sendMessage(e, commandModules.build());
                 } catch(Exception ex) {